% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello  function}
\usage{
hello(lista)
}
\arguments{
\item{lista}{an object of \code{list} class}
}
\value{
\code{list}
}
\description{
hello  function
}
\details{
Function is calculating mean of all vectors in given list.
}
\examples{
hello(list(c(1, 2, 3), c(1, 2, 3), c(12, 34, 56)))
}
\seealso{
\code{\link[base]{debug}}
}
